tests, moved types, renovation
