+++ src/RaggedBooks.Tests/GlobalUsing.cs
@@ -1,0 +1,11 @@
+﻿global using System.Globalization;
+global using Microsoft.Extensions.Logging;
+global using AutoFixture;
+global using AutoFixture.NUnit4;
+global using CsvHelper;
+global using CsvHelper.Configuration;
+
+global using RaggedBooks.Core;
+global using RaggedBooks.Core.Configuration;
+global using RaggedBooks.Core.SemanticSearch;
+global using RaggedBooks.Core.TextExtraction;
