+++ src/RaggedBooks.Tests/OllamaManagerIntegrationTests.cs
@@ -1,8 +1,4 @@
-﻿using Microsoft.Extensions.Logging;
-using RaggedBooks.Core;
-using RaggedBooks.Core.Configuration;
-
-namespace RaggedBooks.Tests;
+﻿namespace RaggedBooks.Tests;
 
 public class OllamaManagerIntegrationTests
 {
